+++ src/library/filter/TriangleFilter.h
@@ -6,19 +6,19 @@
 namespace PR {
 class PR_LIB TriangleFilter : public IFilter {
 public:
-	explicit TriangleFilter(size_t radius = 1)
+	explicit TriangleFilter(int radius = 1)
 		: mRadius(radius)
 	{
 		cache();
 	}
 
-	size_t radius() const override { return mRadius; }
+	int radius() const override { return mRadius; }
 	float evalWeight(float, float) const override;
 
 private:
 	void cache();
 
-	size_t mRadius;
+	int mRadius;
 	std::vector<float> mCache;
 };
 } // namespace PR
